Refactor dependency parser
