+++ src/main/scala/com/johnsnowlabs/nlp/annotators/parser/dep/DependencyParserModel.scala
@@ -22,7 +22,7 @@ class DependencyParserModel(override val uid: String) extends AnnotatorModel[Dep
 
   def getDependencyParsedSentence(sentence: PosTaggedSentence): DependencyParsedSentence = {
     val model = new GreedyTransitionApproach()
-    val dependencyParsedSentence = model.parse(sentence, $(perceptronAsArray))
+    val dependencyParsedSentence = model.predict(sentence, $(perceptronAsArray))
     dependencyParsedSentence
   }
