+++ src/main/scala/com/johnsnowlabs/nlp/annotators/parser/dep/Tagger.scala
@@ -80,10 +80,4 @@ class Tagger(classes: Vector[ClassName], tagDict: Map[Word, ClassNum])  {
       perceptron.toString().split("\\n")
   }
 
-//  def getHarcodedPerceptron: Array[String] = {
-//    val filename = "src/test/resources/models/dep-model-small.txt"
-//    val fileContents = Source.fromFile(filename).mkString
-//    fileContents.split("\\n")
-//  }
-
 }
