+++ src/test/scala/com/johnsnowlabs/nlp/annotators/parser/typdep/TypedDependencyModelTestSpec.scala
@@ -65,7 +65,7 @@ class TypedDependencyModelTestSpec extends FlatSpec {
   }
 
   "A typed dependency parser model with a sentence input" should
-    "predict a labeled relationship between words in the sentence" ignore {
+    "predict a labeled relationship between words in the sentence" in {
     import SparkAccessor.spark.implicits._
 
     val pipeline = new Pipeline()
@@ -92,7 +92,7 @@ class TypedDependencyModelTestSpec extends FlatSpec {
 
 
   "A typed dependency parser model with a document input" should
-    "predict a labeled relationship between words in each sentence" ignore {
+    "predict a labeled relationship between words in each sentence" in {
     import SparkAccessor.spark.implicits._
 
     val pipeline = new Pipeline()
@@ -118,7 +118,7 @@ class TypedDependencyModelTestSpec extends FlatSpec {
   }
 
   "A typed dependency parser model with finisher in its pipeline" should
-    "predict a labeled relationship between words in each sentence" ignore  {
+    "predict a labeled relationship between words in each sentence" in  {
     import SparkAccessor.spark.implicits._
 
     val finisher = new Finisher().setInputCols("labdep")
@@ -147,7 +147,7 @@ class TypedDependencyModelTestSpec extends FlatSpec {
   }
 
   "A typed dependency parser model with an input of more than one row" should
-    "predict a labeled relationship between words in each sentence" ignore  {
+    "predict a labeled relationship between words in each sentence" in  {
     import SparkAccessor.spark.implicits._
 
     val pipeline = new Pipeline()
@@ -198,6 +198,7 @@ class TypedDependencyModelTestSpec extends FlatSpec {
 
     val sentence =
       "The most troublesome report may be the August merchandise trade deficit due out tomorrow"
+    //val sentence = "Book me the morning flight"
     val testDataSet = Seq(sentence).toDS.toDF("text")
     val typedDependencyParserDataFrame = model.transform(testDataSet)
     typedDependencyParserDataFrame.collect()
